created Board.js and BoardList.js
